requirements.txt
numpy
scikit-fuzzy
Pillow
